pyproject.toml
[tool.poetry]
name = "pyaptly"
version = "2.0.0"
description = "Automates the creation and managment of aptly mirrors and snapshots based on yml input files."
authors = ["Jean-Louis Fuchs <jean-louis.fuchs@adfinis.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.mypy]
check_untyped_defs = true

[tool.isort]
profile = "black"

[tool.poetry.scripts]
pyaptly = 'pyaptly.cli:cli'

[tool.flake8]
ignore = [
    # whitespace before ':'
    "E203",
    # too many leading ### in a block comment
    "E266",
    # expected 2 blank lines, found 1 (flake8 and black disagree on what to do
    # with multiple comment blocks at top level)
    "E302",
    # line too long (managed by black)
    "E501",
    # Line break occurred before a binary operator (this is not PEP8 compatible)
    "W503",
]

max-line-length = 88

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "pragma: todo cover",
    "@pytest.mark.skip",
]
omit = []
show_missing = true

[tool.poetry.dependencies]
python = "^3.11"
pretty-dump = {git = "https://github.com/adfinis/freeze"}
pytz = "^2023.3.post1"
pyyaml = "^6.0.1"
toml = "^0.10.2"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
freezegun = "^1.2.2"
hypothesis = "^6.87.1"
testfixtures = "^7.2.0"
mock = "^5.1.0"

pytest = "^7.4.3"
mypy = "^1.7.1"
pdbpp = "^0.10.3"
black = "^23.11.0"
isort = "^5.12.0"

python-lsp-server = "^1.9.0"
python-lsp-black = "^1.3.0"
python-lsp-isort = "^0.1"

flake8 = "^6.1.0"
flake8-bugbear = "^23.12.2"
flake8-debugger = "^4.1.2"
flake8-isort = "^6.1.1"
flake8-docstrings = "^1.7.0"
flake8-string-format = "^0.3.0"
flake8-tuple = "^0.4.1"
flake8-pyproject = "^1.2.3"

types-toml = "^0.10.8.7"
types-pyyaml = "^6.0.12.12"
pytest-coverage = "^0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
